Fix: exhaustive retrieval with parents / tags array
Passing empty arrays to include or exclude tags/parents from retrievals resulted in adding  "tags @> empty array" clauses to SQL queries which is always true--and the NOT clauses were always false.

Fixed here so that no clause is added when an empty array is passed.

Locally tested

To be deployed monday 16.