[core] Fix: splitter issue: nodes with content are always leaf nodes
The chunking algorithm is based on the assumption that the tree induced by a
TokenizedSection must be such that all nodes with content are leaf
nodes.

Therefore, if a section has content and children, in the tokenized
section, the content should be pushed down as a children too and not
remain in the parent node. Whereas the fact that a section has a
prefix should not matter.

We formerly considered whether the section has a prefix, here we fix
so that we consider whethere it has children