[Linter] Allow _, _0, ... , _9 as unused variable names
Linting to avoid unused variables is great. However, there are cases
where unused variables are not avoidable. For instance when using
destructuring, such as in the
following function

```
([scope, _]) => isBuilder(owner) || scope !== "workspace"
```

The following code is IMHO cleaner than not destructuring or adding a
linting ignore -- but does not pass the branch check

While there are multiple solutions, the underscore is a common
convention for "allowed" unused variables

This PR does it