User message passes full name to model if available (#3436)
And username as fallback

username is currently set by the provider and unchangeable by
us (contrarily to first name / last name). In most cases it resembles
the real user's name, but in some it doesn't. The full name is
arguably better when available (i.e. in the case of a google user
pr@dust.tt called Philippe Rolet, it's better that the model thinks
the message comes from 'Philippe Rolet' directly than from 'pr'.

In addition, there are edge cases that causes various issues. Notably,
a user authing with google email marketing@toto.com was
always getting marketing-related answers from the model. We could not
change the username, determined by google as the first part of their
email. This PR allows changing.