[Content Fragment] Migrate deprecated `sourceUrl` fields
Legacy content fragment messages may have a sourceUrl with a private GCS link to extracted textual content, only accessible to Dust so non-downloadable. This PR resets those to a null sourceUrl, to enforce the invariant that sourceUrl is a user-accessible link to an original version of the content--which we do not have for old content fragment.