Fix: prevent editing/deleting already-deleted assistants
Description
---
Deleted assistants can still be accessed in former
conversations. Their details can still be viewed.

At this time, when seeing the details of a deleted assistant:
- there is no visual indication that it was deleted;
- users can still edit them, saving actually revives the assistant;
- users can "delete" them (no-op, confusing).

This PR adresses the 3 as a first effort by preventing edition of
deleted assistants and visually indicating deletion.

We might want a deeper reflexion on what to do on deleted
assistants (maybe a menu with "revive", etc.). This PR can be seen as a
hotfix in the meantime.

Risk
---
No real risk here

Deploy Plan
---
Deploy front