[Dust apps] Datasets - always validate an input of type "string" (#5684)
* [Dust apps] Datasets - accept all entries as string

Description
---
Previously, if an integer was input in a string field of a dataset,
there would be an error; see e.g. [here]()

All that can be typed by the user here is a valid string and should be
accepted as such. This PR takes care of that

Also adds some typing

Risk
---
Low: simple change, restricted to dataset creation

Deploy
---
front

* fix