Remove -_ from sIds
Description
---
Following this [discussion](https://dust4ai.slack.com/archives/C050SM8NSPK/p1719299080025739)

This PR moves sIds to a 62-length charset [A-Za-z0-9]

Note: since 62 is not a power of 2, we'll always have to discard some
bytes. This PR handles it by retrying in unlikely cases where this
causes an issue.

Risk & deploy
---
na
not necessary to backfill IMO