[Vaults] Add groupIds column to agent_configurations
Description
---
Part of # ,  listing only agents the user has access to, by checking
the groupIds columns of agents.

The plan is that user should be in the intersection of all groupIds
here to see the agent. If groupIds is empty then agent is accessible, e.g. if the agent has
no search / table query action then it's not restricted to anybody.

This PR adds the `groupIds` column and properly computes it on agent
creation / edition.

Risk
---
Usual risks from migrations. Here we add a column.

Deploy
---
- run migration
- deploy front