[Global Agents] Return global agent with missing datasource (#7147)
Description
---
Earlier, we had decided to hide agents that could not be used due to a
missing datasource.

Now, we retired "connected" agents, so the only remaining agent
through this code path is `@dust`

This raised an issue because enabling / disabling dust and changing
data it accesses is done via the "Manage" button that is only shown on
the `dust` line of the `default` tab of the assistant listing

If a user disables all datasources, the dust assistant becomes
invisible, and can't be reenabled. Former conversations with it fail,
etc. This happened e.g. [here](https://dust4ai.slack.com/archives/C050SM8NSPK/p1725534664636849)

This PR lets `@dust` show as a disabled agent when it has no
datasources, which fixes the issue.

Risk
---
na, worst case is some agents with a disabled toggle show up which is
really not very scary

Deploy
---
front