[Manage Assistants] Fix: usage sort broken for shared/default tabs (#7486)
Description
---
Fixes https://github.com/dust-tt/tasks/issues/1286

We did not return usage for published or global
agents. Likely because it was not needed when the page was made (there
was no sort button) and then not updated when the sort button was added

Risk
---
Since the cause of not returning usage is not certain, a little risk,
but blast radius is very low. We already did return usage for company
& personal assistants without issues.

Deploy
---
front