[Confluence] Remove mutex lib, use cacheWithRedis with lock
Description
---
After IRL discussion & ensuing thinking:
- remove the `async-mutex` lib (using a mutex library is bad
form in general)
- use cacheWithRedis for the confluence case (see details in PR #7906
for the need for this caching)
- adds a custom lock to cacheWithRedis.

For cacheWithRedis, a lock is legitimate (and needed to adress the
initial issue of #7906), but in order to limit the scope of locking
and avoid spreading its use in our codebase, we use a simple self-made
inline implementation rather than a lib.

Locking operation is quite fast, it shouldn't impact run (but waiting
for the lock could and will).

Implementing the lock could have the added benefit of fixing former unnoticed issues of caching not being used efficiently.

Risks
---
Blast radius: all cacheWithRedis calls => careful. See deploy plan

Deploy plan
---
- local tests on a few cachewithredis calls
- deploy connectors
- deploy front
- monitor deployment
- warn eng runner