[Assistant views] Remove 'assistant-search' & '{conversationId}' views
Description
---
- Removes 'assistant-search' view, which was actually exactly the same
as "list". The comment indicating inactive global agents were returned
was erroneous

- Removes '{conversationId}' view, since:
 - it was actually only used in AssistantSuggestion, not anywhere else;
 - the view didn't actually work, it was ~equivalent to list;
 - the use case it was supposed to enable, viewing a shared
 conversation with somebody else's personal assistant, is already
somewhat working: user can view conversation but not mention assistant
or see its details which is kind of fine.

We could aim in a follow-up PR to reinstate the conversation view
cleanly; that would require more work and is IMO not a priority.

Risks
---
Low, since it's just a view change. Tested locally

Deploy
---
front