[JIT Datasources] Dust apps: restrict access to "conversations" space
Description
---
Fixes https://github.com/dust-tt/tasks/issues/1658

As discussed in shipping JIT datasources, the "conversations" space is
readable by all and as such should not be accessed by API, including
by running a dust app with a datasource block.

This is to prevent an attacker enumerating data source[view]s ids
passing them as config of a datasource block in a dust app

This is particularly important with public dust apps that can be executed
on another workspace with no space permissions checks, but the gating
is also legitimate for private apps.

Only case in which this is allowed is for our packaged apps, via a
system key, in particular "assistant-retrieval-v2" that needs access
to the conversation space

This solution relies on the assumptions that:
- system keys are a good way to distinguish internal calls to our
packaged apps, from calls from API users;
- our packaged apps internal calls cannot be made to pass arbitrary
datasource names.

Risks
---
None, since the conversations space is not used yet so
forbiddenAccessToConv is always false

Deploy
---
front