[Assistant Details] Favorites button correctly disabled while updating
Description
---
Fixes #8933 (again)

Former code relied only on `isUpdatingUserFavorite` to show the button
properly disabled.

There was an issue since the update finishes `before` the mutation of
state fully propagates to agents: first, mutate returns stale data
while revalidating (swr); the `await` does not mean we wait for full
revalidation.
If updating favorites is finished before revalidation has occured,
which is quite likely, then the button will show active with a wrong state.

It's only after revalidation has finished that the button should be
re-enabled. This is what the fix is doing in this PR.

Risks
---
none

Deploy
---
front