[Elasticsearch] Fail on node indexation error
Description
---
Previously, when indexing a node in elasticsearch, we didn't fail, but
just logged an error. This was to test the newly-introduced
infrastructure.

This PR switches to erring, which is the correct behaviour now that
the infra is confirmed stable enough.

Risks
---
Very low, no errors have been witnessed last week
https://app.datadoghq.eu/dashboard/ss3-x6k-9kr/elasticsearch?fromUser=true&refresh_mode=sliding&from_ts=1735572313097&to_ts=1736177113097&live=true

Deploy
---
core