[Login flow] Remove duplicate signup for new users joining by email
Description
---
Fixes #9053

People that are not yet dust users and get invited on a workspace see a "join workspace" page with a signup button
[screen]

When signing up, an account is created for them on auth0, then their user & membership rows are created on dust but they are redirected again to the same "join workspace" page which generates confusion (see issue)

This PR fixes that by redirecting new users directly to the callback after clicking on sign up, which will itself redirect to the workspace welcome page (setup first & last name, I'm a pro, etc.)
[screen]

In doing so, it follows the strong assumption from previous PR #6034 that all new users with pending invitations meant to follow those invitations---and as such already saw the "join workspace" page once.

Risks
---
Impacts login of all new users. While potential for issues is limited, caution is warranted:

+ checked that the "join workspace" page does not change state, we don't miss anything in skipping it and going directly to welcome
+ tested locally
+ 2* reviewers

Deploy
---
front