[Nodes core] Clean gdrive parents - script improvement
Description
---
script is too slow to update parents
if files are under a drive that we sync completely, then we don't need
to update them. It's likely a lot of files are like that.

The script skips those files, and moves concurrency to 32

Risks
---
we risk missing files we should update. Let's (me+reviewer) check the
logic twice.

Deploy
---
na