[Tables API] Fix requirement of mimeType
Description
---
Closes https://github.com/dust-tt/tasks/issues/2097

Mimetype was both required by type system, and forbidden by our API,
because

The logic decided by former issue #8758 and PR #9877 , to allow a user
mimetype and validate it, was not respected.

Given the nature of the table endpoint, it's better that we
- keep the decision of a required mimetype from system key;
- keep setting us ourselves when it's a user--as such preventing the
user from setting the mimetype.

This PR does this, by letting the field optional in types.

Risk
---
standard

Deploy
---
front