[Knowledge - Spreadsheets] Message when user clicks 'raw content'
Description
---
Fixes https://github.com/dust-tt/tasks/issues/2105

Clicking "view raw content" on a spreadsheet was leading to an error
message, see issue

This PR is a temporary fix, since in the knowledge tab and in this
modal, it's not possible to know whether the item is a table or
not---connectors' content nodes do not provide the information.

When the nodes core project is shipped, this distinction can be made
and we can remove "View raw contents" for spreadsheets.

Risk
---
none

Deploy
---
front