[Webcrawler] Skip paused connector directly in activity
Description
---
The webcrawler scheduler does not schedule crawls for paused websites,
but if the website is paused after it was scheduled, the crawl will
start.

E.g. in relocations when we pause all connectors, this happens a lot,
[example here](https://dust4ai.slack.com/archives/C05F84CFP0E/p1742808150399699)

This PR performs the check in the activity.

Risks
---
low

Deploy
---
connectors