fix google ai studio api check
Use a valid model to check Google AI Studio API.
"gemini-pro" is no longer on the list of supported models, so adding an API key for Google AI Studio always fails with the message "Error: "models/gemini-pro is not found for API version v1beta, or is not supported for generateContent.". This commit fixes this by changing the request to a newer model version.