[Attach/Knowledge] Log search queries (#11907)
* [Attach/Knowledge] Log search queries

Description
---
This is to allow understanding what people search, among which whether
they paste document URLs in searches, as was discussed during the
attach project

Risk
---
Too much logs. But since search is debounced, it should be
fine. That'd be #dailyUsers*#dailySearches so likely a few thousands
logs per day

Deploy
---
front

* lint