[MCP Extract] Pass agentsList as arg to constructMAAPrompt (#12705)
Description
---
Preparation for creation of the MCP Extract action, replacing the
legacy one.

The Extract action relies on constructPrompt. In very rare cases
constructPrompt needs to list agents, which creates a circular
dependency when using constructPrompt in an MCP server.

Those rare cases are 1. never seen with an extract action and 2. even
if, irrelevant to the action itself which does not needs the
replacement.

This PR therefore allows avoiding the circular dependency

Risks
---
standard

Deploy
---
front