[Include action] Explicit warning message when reaching limit (#13219)
* [Include action] Explicit warning message when reaching limit

Description
---
Previously, warning was part of the "query" and limited . We observed
that sometimes the model seemed to miss the fact that it hadn't
retrieved everyghing.

This PR extracts the warning as a separate possible result and adapts
the wording to make the situation more explicit to the model.

Risks
---
standard. blast radius: include action. tested locally

Deploy
---
front

* since not until

* review