[Actions v2] Handle blob tool output
Description
---
Raw MCP tool results can be base64-encoded blobs. In that case, we want to turn them into "clean" MCP tool results: storing the data in a file, and stripping the raw base64 data from the returned result.

This was already done for "image" tool results. This PR does it for "resource" tool results that are blobs, that is, whose schema is `EmbeddedResourceSchema` with an inner `BlobResourceContentsSchema`.

At the time, none of our internal tools return a blob, but the goal is to replace tools that currently return results conforming to `ToolGeneratedFileSchema`, with a mime type of `INTERNAL_MIME_TYPES.TOOL_OUTPUT.FILE`. We can then remove this mime type entirely, which will be done in a follow-up PR.

Risks
---
standard, blast radius on image gen

Deploy
---
front