[URL Search] Fix: broken search on notion tables (#14102)
* [URL Search] Fix: broken search on notion tables

Fixes dust-tt/tasks#3400

Replaces https://github.com/dust-tt/dust/pull/14091.

Fixes the issue that when searching an url in assistant builder corresponding to a database, the node id returned was always "notion-[ID]", the one of type "table" --- so no result would be shown if viewType was set to "document".

Notion databases have 2 entries in core:

- one of type "table" starting with "notion-";
- one of type "document" starting with "notion-database-".

We need to pick the right one when searching by URL. We here assume the
user is looking for the table 99% of the time, as the table is also
the "container" of child pages of the database.

* review