Do the actual upsert and nothing else in the Postgres case (#14537)
The Postgres and GCS logic repeat do a couple of things in common, beyond what they do differently:

they update the schema
they do the upsert (that's the part that's different)
they invalidate the schema
Note that in the non-truncate case, the GCS logic happens in sqlite, so we can end up having it happen in both processes.

This change makes it so that 1 and 3 are only donc by the GCS logic (if enabled), to avoid doing it twice, which is causing many redundant calls.