Fix Slack bulk channel selection for workspaces with >1000 channels
Fixes https://github.com/dust-tt/tasks/issues/3842

The issue was that when fetching root nodes for the "Select All" functionality,
we were limited to the default pagination limit of 1000 items. For workspaces
with more channels, this meant not all channels
were loaded.

Solution: Use the infinite scroll hook (useInfinitDataSourceViewContentNodes)
to automatically fetch all pages of root nodes when the component mounts.
This ensures all channels are available for bulk selection operations.