New triggerId column in conversation (#15423)
Add a triggerId column to the conversations table. It’s a FK into the triggers table. But note that it’s optional, as not every conversation is created from a trigger.

Also, get rid of the "triggered" value from ConversationVisibility, since we’ll now be able to tell based on the new triggerId column. Instead, triggered conversation should just have the normal "unlisted" value.

Note that at this time, we won’t yet be using the actual value of triggerId. This will come later.

Note that existing triggered conversations will lose their clock icon, since they don't have a triggerId. We can live with this, as it's still under a flag with low usage.