Skip sending openai encrypted metadata across region (#16035)
OpenAI gives us Reasoning encrypted metadata that we store and send back to them on follow-up messages. The challenge is that when a conversation is started with the US OpenAI endpoint, and continued with the EU endpoint (or vice versa), the encrypted metadata blows up because OpenAI uses different keys.

When we detect this situation, we simply don't send the Reasoning metadata. It may degrade LLM perf a bit, but is the best we can do. Over time, this should become a very rare case as new convos in Dust EU will hit OpenAI EU to begin with.