Decode sqids when viewing raw objects in poke (#16245)
Many times, when seeing Sqids (e.g. msv_z2c3wZmBk0), I need to get the underlying resource ID (in this case 40). This change makes it so that when we view raw objects in Poke using JsonViewer, we automatically do that. It gets added in parentheses after the Sqids, e.g. "msv_z2c3wZmBk0 (40)".