Start refresh loop after file uploads/deletes (#16222)
After a file upload or delete, we want the UI to automatically reflect the change. Before this change, this only happened if you refreshed the page, or if you switched to a different window and came back.

To address this, after an upload/delete, we start a loop where we refresh every 5 seconds for 30 seconds. The assumption is that most changes will be reflected before 30s. We don't want to do this more than necessary, since this adds extra requests for the backend to handle.