[Stop Agent] Fix: don't show "stop all" if there's a single agent
Description
---
Fixes
[this](https://dust4ai.slack.com/archives/C09GAQP6T1R/p1758872777392649)

We observed in rare cases that "Stop all agents" show while there's
only one agent mentioned.

This likely comes from a race condition in the update of generated
messages in the useEffect, which this PR fixes.

The first array check avoids unnecessary calls to setGeneratingMessages so
we keep it.

Risks
---
none

Deploy
---
front