Schema: read all rows at once instead of in batches (#16380)
We had an extremely inefficient way of computing the schema now that files are in GCS.

This is an implementation that made sense when rows were in a table, but completely falls over with the GCS impl. At every iteration of the loop, we end up reading the entire file.

The fix is to read them all at once.