Allow new rows to be synced incrementally (#16795)
We have logic in renderAndUpsertPageFromCache to upsert the row in the DB when the row's page changes. But due to a code ordering issue, this only works for existing rows. For new rows notionPageInDb is not yet set, so we end up skipping that logic.

The fix is to simply move this logic lower down