[Front Sorting] Removes string length as a fuzzy ordering criteria
Description
---

Fixes e.g. claude-4.5-haiku coming before sonnet when we type "cl4",
despite the sorting order for global agents.

Basically, string length here is not a criterion genreral enough that
we'd want to sort generically with it in all fuzzy sorts; this
criterion is left to downstream filters and sorts.

Risk
---
low

Deploy
---
front