Improve how we handle empty table schema (#17891)
The code was not correctly handling the case where:

- the schema was marked as stale, or we don't have one at all AND
- we are not able to compute a new schema.

This should not be common, but can happen if a table is created, but no rows are upserted. It can also happen if we're doing a relocation, and did not copy the GCS buckets.

When we detect this case, we cut short and don't try to save a schema.