Ignore notion linked databases (#18188)
A child database with title "Untitled" is an almost certain sign of a linked database. We ignore those because we will fail to retrieve their content later with two possible errors:

Database with ID xyz is a linked database. Database retrievals do not support linked databases.
Database with ID xyz does not contain any data sources accessible by this API bot.
If we don't, we will end up doing a lot of extra work, and storing bogus entries in notion_databases.

Note that if you create a regular (non-linked) database in Notion without giving it a name, it will be named something like "New database", and not "Untitled". So someone would have to explicitly name a database "Untitled" for us to ignore it, and we can live with that.

Unfortunately Notion does not provide any other reliable way
to detect linked databases when we enumerate blocks.