Avoid setting parents multiple times on the same nodes (#18189)
== Problem description
updateAllParentsFields starts with a clean list of numerical IDs to process (createdOrMovedNotionPageIds/createdOrMovedNotionDatabaseIds), but then getPagesAndDatabasesToUpdate adds all of the descendants to the list.

And then each of the resulting items gets upserted.

When this logic runs after a parentsLastUpdatedAt reset (or in the initial sync), we end up reprocessing the same nodes multiple time.

e.g. in one customer case, I saw 600k 'set parent' calls in core, while they only have a total of 97588 documents. It took 8 hours to complete just the Update Parents phase.

== Fix description
The fix is to special case the scenario where parentsLastUpdatedAt was reset (or it never ran). This happens for the initial sync, and when we reset it via some plugin. In that case, we just process the list that comes in, without adding the descendants.