Ignore DBs that fail validation (#18550)
In most cases, a getParsedDatabase failure means we got a validation error from Notion when trying to fetch the database, commonly because it's either a linked database or it has multiple data sources. In this case, it's better to skip it than end up with a NotionDatabase with many null fields, including parentId.