Don't clear ParentsLastUpdatedAt in update orphaned workflow (#18752)
The call to clearParentsLastUpdatedAt() at the end of updateOrphanedResourcesParentsWorkflow is I believe unnecessary, because the nodes that we update already get an updated lastCreatedOrMovedRunTs, meaning they'll get processed in notionUpdateAllParentsFieldsWorkflow in the next incremental sync.

And setting clearParentsLastUpdatedAt() makes the plugin massively more expensive, because it causes all nodes to be reprocessed. Besides we have another plugin to call it if needed (unstuck syncing nodes).