Exclude tools from search if there is a matching connector (#18968)
If a user has access to a connector, we exclude the matching tool from the search, as that could result in duplicate results.

Note that in theory, the connector and the tool could be exposing a different subset of files, which would not result in duplicates. But this is not common, and we're optimizing for the common case.

I also got rid of searchToolFiles, which was unused and would have needed to be modified accordingly.